+++ examples/server_integration/client/src/lib.rs
@@ -57,7 +57,7 @@ fn update(msg: Msg, model: &mut Model, orders: &mut impl Orders<Msg>) {
 //     View
 // ------ ------
 
-fn view(model: &Model) -> impl View<Msg> {
+fn view(model: &Model) -> impl IntoNodes<Msg> {
     div![
         style! {
             St::FontFamily => "sans-serif";