+++ examples/subscribe/src/lib.rs
@@ -125,7 +125,7 @@ fn update(msg: Msg, model: &mut Model, orders: &mut impl Orders<Msg>) {
 //     View
 // ------ ------
 
-fn view(model: &Model) -> impl View<Msg> {
+fn view(model: &Model) -> impl IntoNodes<Msg> {
     let centered_column = style! {
         St::Display => "flex",
         St::FlexDirection => "column",