+++ src/virtual_dom/view.rs
@@ -1,19 +1,8 @@
-use super::{El, Node};
+use super::{IntoNodes, Node};
 
-pub trait View<Ms> {
+#[deprecated(since = "0.7.0", note = "Use `IntoNodes` instead.")]
+pub trait View<Ms>: IntoNodes<Ms> {
     fn els(self) -> Vec<Node<Ms>>;
-}
-
-impl<Ms> View<Ms> for El<Ms> {
-    fn els(self) -> Vec<Node<Ms>> {
-        vec![Node::Element(self)]
-    }
-}
-
-impl<Ms> View<Ms> for Vec<El<Ms>> {
-    fn els(self) -> Vec<Node<Ms>> {
-        self.into_iter().map(Node::Element).collect()
-    }
 }
 
 impl<Ms> View<Ms> for Node<Ms> {